Alertas para Eliminar y Guardar registros

+++ src/modules/daybook/views/EntryView.vue
@@ -58,6 +58,7 @@
 <script>
 import {defineAsyncComponent } from 'vue'
 import { mapGetters, mapActions } from 'vuex';  //propiedades computadas
+import Swal from 'sweetalert2'
 
 import getDayMonthYear from '../helpers/getDayMonthYear';
 
@@ -114,6 +115,12 @@ export default {
 
         },
         async saveEntry() {
+
+            new Swal({
+                title: 'Espere por Favor',
+                allowOutsideClick: false,
+            })
+            Swal.showLoading()
             
             if( this.entry.id ){
                 // Actualizar
@@ -124,15 +131,33 @@ export default {
                 
                 // redirecTo => entry, param: id
                 this.$router.push({ name: 'entry', params: { id } })
-                
             }
+
+            Swal.fire('Guardado', 'Entrada registrada con éxito', 'success')
             
         },
         
         async onDeleteEntry() {
 
-            await this.deleteEntry( this.entry.id )
-            this.$router.push({ name: 'no-entry' })
+            const { isConfirmed } = await Swal.fire({
+                title: '¿Está seguro?',
+                text: 'Una vez borrado, no se puede eliminar',
+                showDenyButton: true,
+                confirmButtonText: 'Si, estoy seguro'
+            })
+
+            if( isConfirmed ){
+                new Swal({
+                    title: 'Espere por Favor',
+                    allowOutsideClick: false
+                })
+                Swal.showLoading()
+                await this.deleteEntry( this.entry.id )
+                this.$router.push({ name: 'no-entry' })
+
+                Swal.fire('Eliminado', '', 'success')
+            }
+
         }
     },
